refactor and improve routes
